Componentized display when not mobile, Variable Name Change
