+++ sheepUi/pages/index.tsx
@@ -5,6 +5,7 @@ import { sheepData } from '@/utils/data'
 import { AnimatePresence } from 'framer-motion'
 import { useEffect, useState } from 'react'
 import { useSwipeable } from 'react-swipeable'
+import NonMobileDisplay from '../components/NonMobileDisplay'
 import Appbar from '../components/appbar'
 import Page from '../components/page'
 
@@ -29,7 +30,7 @@ import { SheepUpContractAbi } from '../data/SheepUpContractAbi'
 import { SheepUpContractAddress } from '../data/SheepUpContractAddress'
 
 //graph ql
-import { Sheepend } from '../graphql/SheepUpGraph'
+import { GetAllSheepStatus } from '../graphql/SheepUpGraph'
 
 type SheepenedData = {
 	blockNumber: string
@@ -50,13 +51,12 @@ const Index = () => {
 	const [taps, setTaps] = useState<number[]>([])
 	const [ship, setShip] = useState(0)
 
-	async function sheepend() {
+	async function getAllSheepStatus() {
 		console.log('sheepend')
 		try {
-			const data = await Sheepend()
+			const data = await GetAllSheepStatus()
 
 			if (data && data.sheepeneds) {
-				console.log('@@@data.sheepeneds1=', data.sheepeneds)
 				const formattedSheepData = data.sheepeneds.map((sheep: any) => ({
 					...sheep,
 					id: parseInt(sheep.id, 10),
@@ -117,7 +117,7 @@ const Index = () => {
 	}, [])
 
 	useEffect(() => {
-		sheepend()
+		GetAllSheepStatus()
 	}, [])
 
 	//send tx
@@ -166,27 +166,10 @@ const Index = () => {
 	const shipTapNumber = tapStamina ? Number(tapStamina) : 0
 	const pointNumber = point ? Number(point) : 0
 
-	const NotSpDisplay = () => {
-		return (
-			<div className='hidden sm:flex flex-col items-center justify-center bg-blue h-screen'>
-				<img src='/images/sheep.svg' alt='Sheep Icon' className='w-18 h-18' />
-				<p className='p-2 font-bold text-4xl text-white'>Sheep Up</p>
-				<p className='p-2 font-bold text-md text-white'>
-					An onchain Sheeping (Sheep * ship it) game
-				</p>
-				<div className='bg-gray rounded-md p-4 mt-8'>
-					<p className='text-black font-bold'>
-						Sheep Up is only on mobile. Visit on your phone to play.
-					</p>
-				</div>
-			</div>
-		)
-	}
-
 	if (!address) {
 		return (
 			<>
-				<NotSpDisplay />
+				<NonMobileDisplay />
 				<div className='sm:hidden absolute top-0 left-0 h-screen w-screen bg-stone-100'>
 					<Login />
 				</div>
@@ -196,7 +179,7 @@ const Index = () => {
 	if (!balance) {
 		return (
 			<>
-				<NotSpDisplay />
+				<NonMobileDisplay />
 				<div className='sm:hidden absolute top-0 left-0 h-screen w-screen bg-stone-100'>
 					<GetEth address={address} />
 				</div>
@@ -206,7 +189,7 @@ const Index = () => {
 
 	return (
 		<>
-			<NotSpDisplay />
+			<NonMobileDisplay />
 			<div className='sm:hidden'>
 				<Appbar
 					furAmount={pointNumber}
